Move SlugInput to admin component
