+++ front/app/components/admin/NewSlugInput/index.tsx
